requirements.txt
djangorestframework
validate_email
markdown
django
gunicorn
django-cors-headers
jsonfield
